requirements.txt
Flask
transformers
psycopg2-binary
